+++ Src/Inventory/SupplyNest.Inventory.Api/Domain/ValueObjects/InventoryQuantity.cs
@@ -7,15 +7,15 @@ namespace SupplyNest.Inventory.Api.Domain.ValueObjects
 {
     public record InventoryQuantity
     {
-        public int Value { get; init; }
+        public long Value { get; init; }
 
 
-        private InventoryQuantity(int value)
+        private InventoryQuantity(long value)
         {
             Value = value;
         }
     
-        public static InventoryQuantity FromInt(int value) =>
+        public static InventoryQuantity FromInt(long value) =>
             new(value);
 
         public static InventoryQuantity operator +(InventoryQuantity a, InventoryQuantity b) =>
@@ -24,6 +24,6 @@ namespace SupplyNest.Inventory.Api.Domain.ValueObjects
         public static InventoryQuantity operator -(InventoryQuantity a, InventoryQuantity b) =>
             new(a.Value - b.Value);
 
-        public static implicit operator int(InventoryQuantity quantity) => quantity.Value;
+        public static implicit operator long(InventoryQuantity quantity) => quantity.Value;
     }
 }
